Rename local functions and embed in anonymous namespace
The templated function vtkMapperScalarToTextureCoordinate was defined
within the implementation file for vtkScalarsToColorsPainter. An
identical function with the same name was also defined in the
implementation file for vtkMapper. As is done with multiply-defined
templated functions, the linker was selecting only one of these
definitions for use by both vtkScalarsToColors and vtkMapper,
which is obviously not desirable behavior, but which turned out
no to cause any problems because the definitions were the same.

Solved this problem by placing both function definitions in anonymous
namespaces. Also, reduce the size of the names of these functions to
exclude the class name prefix.

Change-Id: Id80f13d68183b3045f0a99fb5b801c477959cde4