Change clamping range for texture coordinates
When InterpolateScalarsBeforeMapping is on, vtkMapper and
vtkScalarsToColorsPainter would clamp texture coordinates to the range
[0, 1]. This lead to incorrect results if the range of the scalar
values associated with points that define a primitive were not
completely contained within the range of the color lookup table used
to color that primitive, e.g., a quad with scalar range [0,3] colored
by a color map with range [1,2], and the option
InterpolateScalarsBeforeMapping is on. In this case, the texture
coordinates should be in the range [-1, 2], not clamped to [0, 1].

Unfortunately, removing the clamping completely leads to another
problem. It turns out that some implementations of OpenGL do not
handle arbitrary texture coordinates of about 1000 or higher
correctly. To handle this, clamping has been retained, but to the
range [-1000, 1000], and a warning that colors may not be correct is
emitted when this clamping occurs. Hopefully this covers most use
cases.

Change-Id: I77a7d203a54a5bb66ca5005a747510504929d83c