Fixed errors exposed by changes to clamping of texture coordinates
Told the mapper to use the lookup table's range. Without doing that,
the mapper was using the range [0, 1] for the color map range instead
of approximately [-250, 250]. The somewhat correct (but wrong)
appearance of the original baselines was a side-effect of the
incorrect clamping. Baselines have been updated to reflect the
corrected color mapping.

Change-Id: Ibd78de5464c542130a381d0ad53c21f49b17ff1c